# pastel360-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PASTEL 360º API RESTFUL",
        "description": "API para gerenciamento de clientes, produtos e pedidos",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Lista todos os pedidos",
                "operationId": "2cd7e03021e72d9d799c12dfe179c865",
                "responses": {
                    "200": {
                        "description": "Lista de pedidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Cria um novo pedido",
                "operationId": "d690d6834318e9b064df441e854de8b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "client_id",
                                    "quantity",
                                    "unit_value",
                                    "total_value",
                                    "status"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "client_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "unit_value": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1500
                                    },
                                    "total_value": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 3000
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "approved",
                                            "canceled",
                                            "delivered"
                                        ],
                                        "example": "pending"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Entregar de manhã"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validação falhou"
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Obtém um pedido específico",
                "operationId": "c29621029e70f65247909d3e09031118",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pedido encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Atualiza um pedido",
                "operationId": "6a7029f5510ca08e28a2d58a5fe19439",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "client_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "unit_value": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1500
                                    },
                                    "total_value": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 3000
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "approved",
                                            "canceled",
                                            "delivered"
                                        ],
                                        "example": "approved"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Entregar de manhã"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pedido atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    },
                    "422": {
                        "description": "Validação falhou"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Exclui um pedido",
                "operationId": "303a7b797887b66c7790c4c187b54790",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Pedido excluído com sucesso"
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Client": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "José Romano"
                    },
                    "mail": {
                        "type": "string",
                        "format": "email",
                        "example": "jromano@teste.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "11999999999"
                    },
                    "birthdate": {
                        "type": "string",
                        "format": "date",
                        "example": "1996-01-01"
                    },
                    "place": {
                        "type": "string",
                        "example": "Rua Principal"
                    },
                    "number": {
                        "type": "string",
                        "example": "123"
                    },
                    "zipcode": {
                        "type": "string",
                        "example": "12345678"
                    },
                    "district": {
                        "type": "string",
                        "example": "Centro"
                    },
                    "complement": {
                        "type": "string",
                        "example": "Apto 45"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Order": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "client_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "unit_value": {
                        "type": "number",
                        "format": "float",
                        "example": 1500
                    },
                    "total_value": {
                        "type": "number",
                        "format": "float",
                        "example": 3000
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Entregar de manhã"
                    },
                    "client": {
                        "$ref": "#/components/schemas/Client"
                    },
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Notebook Dell"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 2999.99
                    },
                    "photo": {
                        "type": "string",
                        "example": "notebook.jpg"
                    },
                    "stock": {
                        "type": "integer",
                        "example": 50
                    },
                    "sku": {
                        "type": "string",
                        "example": "DELL-NB-001"
                    },
                    "enable": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}